Project Update
Updated the project to just a main file with the alternate version as the main.
Updated the game to include score keeping and to display score after each round.
Updated game to determine overall winner when the player quits.
Updated readme to reflect the changes and to cite the source for the version

No issues to track